pyproject.toml
[project]
name = "family-agent"
version = "0.1.0"
description = "Family Management System API Service"
authors = []
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.2",
    "pydantic-settings>=2.0.0",
    "line-bot-sdk>=3.16.0",
    "openai>=1.70.0",
    "openai-agents[litellm]>=0.0.8",
    "firecrawl-py",
    "mem0ai>=0.1.114",
    "google-maps-places>=0.2.2",
    "mcp>=1.11.0",
    "anthropic>=0.59.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10",
]
requires-python = ">=3.11"

[project.optional-dependencies]
test = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.23.2",
]
dev = [
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
]
